+++ src/ProxyManager.php
@@ -2,27 +2,23 @@
 
 namespace Monooso\Unobserve;
 
-use Illuminate\Contracts\Container\Container;
+use Illuminate\Container\Container;
 
 class ProxyManager
 {
-    /** @var Container */
-    private $app;
-
-    public function __construct(Container $app)
+    public function __construct(private Container $app)
     {
-        $this->app = $app;
     }
 
-    public function register($target, array $events)
+    public function register(object $target, array $events): void
     {
         $proxy = $this->app->make(Proxy::class, ['target' => $target, 'events' => $events]);
 
         $this->app->instance(get_class($target), $proxy);
     }
 
-    public function unregister($target)
+    public function unregister(string $targetClass): void
     {
-        $this->app->instance(get_class($target), $target);
+        $this->app->forgetInstance($targetClass);
     }
 }