+++ tests/CanMuteTest.php
@@ -3,12 +3,13 @@
 namespace Monooso\Unobserve\Tests;
 
 use Monooso\Unobserve\CanMute;
+use Monooso\Unobserve\Proxy;
 use Orchestra\Testbench\TestCase;
 
 class CanMuteTest extends TestCase
 {
     /** @test */
-    public function it_mutes_an_array_of_events()
+    public function it_mutes_an_array_of_events(): void
     {
         CanMuteTarget::mute(['cloaked']);
 
@@ -19,7 +20,7 @@ class CanMuteTest extends TestCase
     }
 
     /** @test */
-    public function it_mutes_a_single_event()
+    public function it_mutes_a_single_event(): void
     {
         CanMuteTarget::mute('cloaked');
 
@@ -30,7 +31,7 @@ class CanMuteTest extends TestCase
     }
 
     /** @test */
-    public function it_mutes_all_events()
+    public function it_mutes_all_events(): void
     {
         CanMuteTarget::mute();
 
@@ -41,7 +42,7 @@ class CanMuteTest extends TestCase
     }
 
     /** @test */
-    public function it_unmutes_all_events()
+    public function it_unmutes_all_events(): void
     {
         CanMuteTarget::mute();
         CanMuteTarget::unmute();
@@ -51,19 +52,59 @@ class CanMuteTest extends TestCase
         $this->assertSame('cloaked', $target->cloaked());
         $this->assertSame('uncloaked', $target->uncloaked());
     }
+
+    /** @test */
+    public function it_mutes_class_with_constructor_injection(): void
+    {
+        WithConstructorInjection::mute();
+
+        $target = resolve(WithConstructorInjection::class);
+
+        $this->assertInstanceOf(Proxy::class, $target);
+    }
+
+    /** @test */
+    public function it_resolves_to_proxy_on_mute(): void
+    {
+        CanMuteTarget::mute();
+
+        $target = resolve(CanMuteTarget::class);
+
+        $this->assertInstanceOf(Proxy::class, $target);
+    }
+
+    /** @test */
+    public function it_resolves_back_to_class_on_unmute(): void
+    {
+        CanMuteTarget::mute();
+        CanMuteTarget::unmute();
+
+        $target = resolve(CanMuteTarget::class);
+
+        $this->assertInstanceOf(CanMuteTarget::class, $target);
+    }
 }
 
 class CanMuteTarget
 {
     use CanMute;
 
-    public function cloaked()
+    public function cloaked(): string
     {
         return 'cloaked';
     }
 
-    public function uncloaked()
+    public function uncloaked(): string
     {
         return 'uncloaked';
+    }
+}
+
+class WithConstructorInjection
+{
+    use CanMute;
+
+    public function __construct(private CanMuteTarget $injection)
+    {
     }
 }