+++ tests/ProxyManagerTest.php
@@ -9,7 +9,7 @@ use Orchestra\Testbench\TestCase;
 class ProxyManagerTest extends TestCase
 {
     /** @test */
-    public function it_registers_a_proxy_with_the_service_container()
+    public function it_registers_a_proxy_with_the_service_container(): void
     {
         $app = $this->resolveApplication();
 
@@ -20,12 +20,12 @@ class ProxyManagerTest extends TestCase
     }
 
     /** @test */
-    public function it_unregisters_a_proxy()
+    public function it_unregisters_a_proxy(): void
     {
         $app = $this->resolveApplication();
 
         $manager = new ProxyManager($app);
-        $manager->unregister(new ProxyManagerTarget);
+        $manager->unregister(ProxyManagerTarget::class);
 
         $this->assertInstanceOf(ProxyManagerTarget::class, $app->make(ProxyManagerTarget::class));
     }